add prototype for the GribJumpSource based on SimpleFieldList
Support for to_xarray, robust selection, efficient loading of subsets, tests, etc. still missing.

This approach is inspired by FieldlistFromDicts and GribFieldListInMemory and I suspect
this approach is most in line with how similar problems have been dealt with in earthkit-data
by other contributors.

The main issues I encountered are the following:

The extracted output of GribJump is not an actual Field but an array with raw,
extracted values from different grid cells in the field. Therefore, most of the
existing abstractions and utilities do not fit this scenario.  After trying out
a few different things myself, I found the "list-of-dicts" source which
theoretically also supports non-field values (numpy arrays with arbitrary
metadata in dictionary form). I am still a bit concerned that breaking
abstractions here in our case could be more confusing and bug-prone in the
future.

Inspired by GribFieldListInMemory, the FieldExtractList wraps the methods of
its superclass SimpleFieldList that make it an iterator to lazily load the data
using gribjump once it's actually needed. However, this might be a bit brittle
as currently implemented and could be improved.