assign grid index to each value in to_xarray
When users create an xarray dataset for the values retrieved via GribJump, we
think it is important for users to explicitly attach the indices from the
original grid to each returned value.

Unfortunately, this is currently not easy to do directly via the UserMetadata /
UserGeomery, which is why we manually add this information in
FieldExtractList.to_xarray for now. This should be cleaned up and likely should
be natively supported in earthkit-data. Also, although returning the index
in the flattened, original grid works for now, we likely want to rather
include the origin (x, y) index and ultimately (lat, lon) coordinates instead
(or additionaly).

The change in pygribjump that enables this lives in [this
PR](https://github.com/ecmwf/gribjump/pull/60).